removed unused code, added comments and moved type defination to types.h
